Changed to /firephone SymLink and more...

+++ event/unplug.sh
@@ -1,3 +1,10 @@
 #!/bin/bash
-killall FireTime
-/User/Documents/FireCloud/cron/deviceinfo.sh
+
+if [ "$(activator current-app)" == "pro.firepanther.FireTime" ]; then
+	killall FireTime
+	activator send libactivator.system.sleepbutton
+fi
+
+activator send switch-on.com.a3tweaks.switch.autolock
+
+/firephone/cron/deviceinfo.sh
